event_service/pyproject.toml
[tool.poetry]
name ="event_service"
version = "0.1.0"
description = ""
authors = [
    "Nhsdkk <38581029+Nhsdkk@users.noreply.github.com>",
    "ZixMai <149195961+ZixMai@users.noreply.github.com.>"
]

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.62.1"
grpcio-tools = "^1.62.1"
types-protobuf = "^4.24.0.20240311"
grpc-interceptor = "0.15.4"
requests = "^2.31.0"
protoletariat = "^3.2.19"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.1"
errors-package = {path = "../shared/errors_package"}
db-package = {path = "../shared/db_package"}
utils-package = {path = "../shared/utils_package"}

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
flake8-docstrings = "^1.7.0"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
